requirements.txt
Flask>=2.2
requests>=2.31.0
cloudscraper>=1.2.71
beautifulsoup4>=4.12.3
lxml>=5.2.2
python-dotenv>=1.0.0
Jinja2>=3.1.3
httpx[http2]>=0.27.0
feedparser>=6.0.10
typer>=0.12.3
waitress>=3.0.0
PyYAML>=6.0.1
psycopg[binary]>=3.1.18
Pillow>=10.3.0
ImageHash>=4.3.1
python-aliexpress-api>=2.0.0
warcio>=1.7.4
pywb>=2.7.6
duckduckgo-search>=6.3.0
ddgs>=9.5.0
